Option Additions
Addition of 'event_custom_css_class' option so that one may customize
the default css class that is to be attached to an event.

Addition of 'data' property in 'req_ajax' option. A user may add
additional information that is to be passed along to the handler along
with 'month' and 'year' which already passed.